Small refactor to consider the xml file as a repo instead of an input file
