+++ README.md
@@ -1,5 +1,5 @@
-##Tech test for bodas.net
+## Tech test for bodas.net
 ### Steps to execute:
 - Clone the repository
 - `composer install`
-- `php bin/console.php bodas:obtain-users <dataFile.xml>`
+- `php bin/console.php bodas:obtain-users` (dataFile.xml must be placed in root dir)
