+++ src/Command/ObtainCsvFileCommand.php
@@ -4,11 +4,11 @@
 namespace App\Command;
 
 
-use App\Application\GenerateClientsCsvListService;
+use App\Application\Service\GenerateClientsCsvListService;
 use App\Infrastructure\Domain\Repository\JsonClientRepository;
+use App\Infrastructure\Domain\Repository\XmlClientRepository;
 use GuzzleHttp\Client;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -29,19 +29,15 @@ class ObtainCsvFileCommand extends Command
 
     protected function configure(): void
     {
-        $this->addArgument('xmlFile', InputArgument::REQUIRED);
-
         $this->generateClientsCsvListService = new GenerateClientsCsvListService(
-            new JsonClientRepository(new Client())
+            new JsonClientRepository(new Client()),
+            new XmlClientRepository()
         );
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $filename = $input->getArgument('xmlFile');
-        $xmlClientData = file_get_contents($filename);
-
-        $clientArray = $this->generateClientsCsvListService->execute($xmlClientData);
+        $clientArray = $this->generateClientsCsvListService->execute();
 
         $this->storeResultToCsv($clientArray);
     }
